Enhance footer component with new 'Manage' button and update button styles for better visibility. Refactor header implementation across multiple pages for consistency and improved user experience.
